twitter_clone_backend/pyproject.toml
[tool.poetry]
name = "twitter-clone-backend"
version = "0.1.0"
description = ""
authors = ["Shawaal Nadeem <110089331+Shawaal-Nadeem@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.21"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
